treat core as a special collection
